requirements.txt
# Core dependencies
chromadb>=0.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
httpx>=0.25.0

# Google Gemini AI
google-generativeai>=0.3.0

# Optional: OpenAI (if you want to switch back)
# openai>=1.0.0

# Web framework (optional, for API)
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.0.0

# Data analysis (optional, for scripts)
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
ipykernel>=6.25.0

